Move judges endpoint to DashboardController to since UserService requires MongoDb instance
