+++ api/Controllers/UsersController.cs
@@ -1,16 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using FluentValidation;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using Scv.Api.Helpers.Extensions;
 using Scv.Api.Infrastructure.Authorization;
 using Scv.Api.Models.AccessControlManagement;
 using Scv.Api.Services;
 using Scv.Db.Models;
-using PCSSModels = PCSSCommon.Models;
 
 namespace Scv.Api.Controllers;
 
@@ -31,27 +26,5 @@ public class UsersController(
     public override Task<IActionResult> GetAll()
     {
         return base.GetAll();
-    }
-
-    [HttpGet]
-    [Route("judges")]
-    public async Task<IActionResult> GetJudges()
-    {
-        if (!this.User.CanViewOthersSchedule())
-        {
-            return Ok(new List<PCSSModels.PersonSearchItem>());
-        }
-
-        try
-        {
-            var result = await this.Service.GetJudges();
-            return Ok(result);
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine("Error retrieving judges.");
-            Console.WriteLine(ex);
-            return Ok(new List<PCSSModels.PersonSearchItem>());
-        }
     }
 }
